requirements.txt
fdb
requests
pyyaml
-e .
